Make mini.statusline dev information bold
The contrast of the dev info text to the background is low (3.59:1),
which fails WCAG AA standard. This information is inactive, so failing
the standard is acceptable per the README. However, it is still quite
hard to read due to the low contrast.

By making the dev info bold, the text becomes much easier to read. This
appears to be the behaviour in the README screenshot, which features
lualine. This patch would make the mini.statusline behaviour consistent
with lualine.